Chorros de cambios para Meet Me.

+++ src/platform_resources.py
@@ -85,21 +85,23 @@ class AzureResourcer():
                 print(f"Blob uploaded: {account}, {container};\n\t{blob}")
         
         
-    def download_storage_blob(self, file, blob, container, account=None): 
+    def download_storage_blob(self, file, blob, container, account=None, verbose=0): 
         if account is None: 
             account = self.get_storage()
         
         the_url = f"https://{account}.blob.core.windows.net"
         service = BlobServiceClient(the_url, credential=self.env.credential)
 
-        if Path(file).is_file(): 
-            remove(file)
-            
         the_blob = service.get_blob_client(container, blob)
-        with open(file, 'wb') as _b: 
-            blob_data = the_blob.download_blob()
-            blob_data.readinto(_b)
-    
+        if the_blob.exists(): 
+            if Path(file).is_file(): 
+                remove(file)
+            with open(file, 'wb') as _b: 
+                blob_data = the_blob.download_blob()
+                blob_data.readinto(_b)
+        else:
+            print(f"\tReading blob: {account}, {container};\n\t{blob}")
+
